requirements.txt
psutil==5.9.5
matplotlib==3.7.2
netifaces==0.11.0
speedtest-cli==2.1.3